firebase like update fix

+++ TwitterClone/ViewModel/FeedViewModel.swift
@@ -10,7 +10,6 @@ import SwiftUI
 class FeedViewModel: ObservableObject {
     @Published var tweets = [Tweet]()
     
-    
     init() {
         fetchTweets()
     }
@@ -26,10 +25,14 @@ class FeedViewModel: ObservableObject {
         
         guard let uid = AuthViewModel.shared.userSession?.uid else { return }
         
-        COLLECTION_TWEETS.document(tweet.id).updateData(["likes": tweet.likes + 1]) { _ in
-            COLLECTION_TWEETS.document(tweet.id).collection("tweet-likes").document(uid).setData([:]) { _ in
-                COLLECTION_USERS.document(uid).collection("user-likes").document(tweet.id).setData([:]) { _ in
-                    
+        COLLECTION_TWEETS.document(tweet.id).getDocument { snapshot, _ in
+            guard let data = snapshot?.data() else { return }
+            let tempTweet = Tweet(dictionary: data)
+            COLLECTION_TWEETS.document(tweet.id).updateData(["likes": tempTweet.likes + 1]) { _ in
+                COLLECTION_TWEETS.document(tweet.id).collection("tweet-likes").document(uid).setData([:]) { _ in
+                    COLLECTION_USERS.document(uid).collection("user-likes").document(tweet.id).setData([:]) { _ in
+                        
+                    }
                 }
             }
         }
@@ -39,10 +42,13 @@ class FeedViewModel: ObservableObject {
         
         guard let uid = AuthViewModel.shared.userSession?.uid else { return }
         
-        COLLECTION_TWEETS.document(tweet.id).updateData(["likes": tweet.likes - 1]) { _ in
-            COLLECTION_TWEETS.document(tweet.id).collection("tweet-likes").document(uid).setData([:]) { _ in
-                COLLECTION_USERS.document(uid).collection("user-likes").document(tweet.id).delete() { _ in
-                
+        COLLECTION_TWEETS.document(tweet.id).getDocument { snapshot, _ in
+            guard let data = snapshot?.data() else { return }
+            let tempTweet = Tweet(dictionary: data)
+            COLLECTION_TWEETS.document(tweet.id).updateData(["likes": tempTweet.likes - 1]) { _ in
+                COLLECTION_TWEETS.document(tweet.id).collection("tweet-likes").document(uid).delete() { _ in
+                    COLLECTION_USERS.document(uid).collection("user-likes").document(tweet.id).delete() { _ in
+                    }
                 }
             }
         }
